requirements.txt
terminaltables
pytorch_pfn_extras
fastargs
matplotlib
sklearn
imgcat
pandas
assertpy
tqdm
psutil
webdataset
torchmetrics